Move all name validity-checking into is_valid_name(). Call is_valid_name() the once; don't split into multiple parts.
In retrospect the alternating iteration mentioned in the last commit doesn't seem to serve much purpose beyond perhaps not needing to iterate over the name twice? In any case, we need to do it only once so we can tell where the end is (so you can't have _ at the end of a name).